% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_lables}
\alias{add_lables}
\title{Add labels to  the first column of a data frame}
\usage{
add_lables(x, value)
}
\arguments{
\item{x}{a data frame.}

\item{value}{a data frame contain the column of term, code, and label.}
}
\value{
a data frame.
}
\description{
Add labels to  the first column of a data frame
}
