% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-aSAH.R
\docType{data}
\name{aSAH}
\alias{aSAH}
\title{Subarachnoid hemorrhage data}
\format{
A data.frame containing 113 observations of 7 variables.
}
\source{
Natacha Turck, Laszlo Vutskits, Paola Sanchez-Pena, Xavier Robin,
Alexandre Hainard, Marianne Gex-Fabry, Catherine Fouda, Hadiji Bassem,
Markus Mueller, Frédérique Lisacek, Louis Puybasset and Jean-Charles Sanchez
(2010) “A multiparameter panel method for outcome prediction following
aneurysmal subarachnoid hemorrhage”. Intensive Care Medicine 36(1), 107–115.
DOI: doi: 10.1007/s00134-009-1641-y.
}
\usage{
aSAH
}
\description{
This dataset summarizes several clinical and one laboratory
variable of 113 patients with an aneurysmal subarachnoid hemorrhage.
}
\keyword{datasets}
