% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing.R
\name{gg_missing}
\alias{gg_missing}
\title{Plots for missing with ggplot2}
\usage{
gg_missing(
  data,
  percent.bar = TRUE,
  percent.miss = FALSE,
  show.all = FALSE,
  decreasing = TRUE,
  detail.type = TRUE,
  add.var.miss = TRUE,
  digits = 1,
  language = NULL,
  bar.color = NULL,
  bar.width = 0.65,
  miss.color = NULL,
  font.family = NULL,
  font.size = NULL,
  tag.levels = "A",
  plot.width = c(0.5, 0.5),
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{percent.bar}{a logical for bar, show percentage or frequency.}

\item{percent.miss}{a logical for missing plot, show percentage or frequency.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{decreasing}{a logical. Should the sort order be increasing or decreasing?
if is NULL, not sorted.}

\item{detail.type}{detail type.}

\item{add.var.miss}{a logical, indicate whether to show missing count of each variable.}

\item{digits}{digits for missing percent, default 1.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{bar.color}{color for bar plot.}

\item{bar.width}{width for bar.}

\item{miss.color}{color for missing plot.}

\item{font.family}{font family, default 'serif'.}

\item{font.size}{font size, default 12.}

\item{tag.levels}{a character vector defining the enumeration format to use at
each level.  It can also be a list containing character vectors defining
arbitrary tag sequences. If any element in the list is a scalar and one of
'a', 'A', '(a)', '(A)', '[a]', '[A]', '(1)'  or '[1]', this level will be
expanded to the expected sequence.}

\item{plot.width}{the relative widths and heights of each column and row in
the grid. Will get repeated to match the dimensions of the grid.}

\item{...}{further arguments pass to \link{gg_theme_sci} function.}
}
\value{
a ggplot.
}
\description{
Plots for missing with ggplot2
}
\examples{
# Basic
gg_missing(lung)

# set tag
gg_missing(lung, tag.levels = c("A: Missing bar", "B: Missing pattern"))
}
