% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing-bar.R
\name{gg_missing_bar}
\alias{gg_missing_bar}
\title{Plot a bar for missing with ggplot2}
\usage{
gg_missing_bar(
  data,
  percent = TRUE,
  show.all = FALSE,
  decreasing = TRUE,
  detail.type = TRUE,
  digits = 1,
  language = c("en", "zh"),
  bar.color = NULL,
  bar.width = 0.65,
  font.family = "serif",
  font.size = 12,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{percent}{a logical, show percentage or frequency.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{decreasing}{a logical. Should the sort order be increasing or decreasing?
if is NULL, not sorted.}

\item{detail.type}{detail type.}

\item{digits}{digits for missing percent, defualt 2.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{bar.color}{color for bar.}

\item{bar.width}{width for bar.}

\item{font.family}{font family.}

\item{font.size}{font size.}

\item{...}{further arguments pass to \link{gg_theme_sci} function.}
}
\value{
a ggplot.
}
\description{
Plot a bar for missing with ggplot2
}
\examples{
# Basic example
gg_missing_bar(lung)

# Show all variable
gg_missing_bar(lung, show.all = TRUE)

# Show missing count
gg_missing_bar(lung, show.all = TRUE, percent = FALSE)
}
