% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{global_option_fontsize}
\alias{global_option_fontsize}
\title{Set global font size}
\usage{
global_option_fontsize(font.size = NULL)
}
\arguments{
\item{font.size}{font.size}
}
\description{
Set global font size
}
