% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-global-options.R
\name{global_option_labels}
\alias{global_option_labels}
\title{Set global labels}
\usage{
global_option_labels(labels = NULL)
}
\arguments{
\item{labels}{a data frame contain the column of term, code, and label.}
}
\description{
Set global labels
}
