% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{global_option_languange}
\alias{global_option_languange}
\title{Set global languange}
\usage{
global_option_languange(language = NULL)
}
\arguments{
\item{language}{lauage}
}
\description{
Set global languange
}
