% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-missing.R
\name{identify_missing}
\alias{identify_missing}
\title{Missing value analysis}
\usage{
identify_missing(
  data,
  show.all = FALSE,
  digits = 2,
  language = c("en", "zh"),
  table.number = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{digits}{digits for missing percent, defualt 2.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{table.number}{table number.}

\item{...}{unused arguments.}
}
\value{
a data frame.
}
\description{
Missing value analysis
}
