% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_missing_pattern.R
\name{identify_missing_pattern}
\alias{identify_missing_pattern}
\title{Identify missing pattern}
\usage{
identify_missing_pattern(
  data,
  show.all = FALSE,
  decreasing = TRUE,
  language = c("en", "zh"),
  table.number = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{decreasing}{a logical. Should the sort order be increasing or decreasing?
if is NULL, not sorted.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{table.number}{table number.}

\item{...}{unused arguments.}
}
\value{
a data frame.
}
\description{
Identify missing pattern
}
\examples{
# Basic example
identify_missing_pattern(lung)

# Show all variables
identify_missing_pattern(lung, show.all = TRUE)
}
