% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\alias{is_outlier}
\alias{is_extreme}
\title{Identify univariate outliers}
\usage{
identify_outliers(data, group = NULL, varnames = NULL, language = NULL)

is_outlier(x, coef = 1.5)

is_extreme(x)
}
\arguments{
\item{data}{a data frame.}

\item{group}{group variable name.}

\item{varnames}{numeric variable names.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{x}{a numeric vector.}

\item{coef}{coefficient specifying how far the outlier should be from the edge
of their box. Possible values are 1.5 (for outlier) and 3 (for extreme points
only). Default is 1.5}
}
\value{
identify_outliers() returns a data frame.

is_outlier() and is_extreme() return logical vectors.
}
\description{
Detect outliers using boxplot methods. Boxplots are a popular and an
easy method for identifying outliers. There are two categories of outlier:
(1) outliers and (2) extreme points.

Values above Q3 + 1.5xIQR or below Q1 - 1.5xIQR are considered as outliers.
Values above Q3 + 3xIQR or below Q1 - 3xIQR are considered as extreme points
(or extreme outliers).

Q1 and Q3 are the first and third quartile, respectively. IQR is the
interquartile range (IQR = Q3 - Q1).

Generally speaking, data points that are labelled outliers in boxplots are not
considered as troublesome as those considered extreme points and might even
be ignored. Note that, any NA and NaN are automatically removed before the
quantiles are computed.
}
