% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lung.R
\docType{data}
\name{lung.labels}
\alias{lung.labels}
\title{Labels for data of lung}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 18 rows and 3 columns.
}
\usage{
lung.labels
}
\description{
Labels for data of lung
}
\keyword{datasets}
