% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lung.R
\docType{data}
\name{lung}
\alias{lung}
\title{NCCTG lung cancer data}
\format{
A data.frame containing 228 observations of 10 variables.
}
\source{
Loprinzi CL. Laurie JA. Wieand HS. Krook JE. Novotny PJ. Kugler JW.
Bartel J. Law M. Bateman M. Klatt NE. et al. Prospective evaluation of
prognostic variables from patient-completed questionnaires. North Central
Cancer Treatment Group. Journal of Clinical Oncology. 12(3):601-7, 1994.
}
\usage{
lung
}
\description{
Survival in patients with advanced lung cancer from the North
Central Cancer Treatment Group. Performance scores rate how well the patient
can perform usual daily activities.
}
\keyword{datasets}
