% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-missing.R
\name{print.srp.miss}
\alias{print.srp.miss}
\title{Print class of 'srp.miss'}
\usage{
\method{print}{srp.miss}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp.miss'
}
\keyword{internal}
