% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_missing_pattern.R
\name{print.srp.mp}
\alias{print.srp.mp}
\title{Print class of 'srp.mp'}
\usage{
\method{print}{srp.mp}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp.mp'
}
\keyword{internal}
