% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-outliers.R
\name{print.srp.outlier}
\alias{print.srp.outlier}
\title{Print class of 'srp.outlier'}
\usage{
\method{print}{srp.outlier}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp.outlier'
}
\keyword{internal}
