% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{print.srp.overview}
\alias{print.srp.overview}
\title{Print class of 'srp.overview'}
\usage{
\method{print}{srp.overview}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp.overview'
}
\keyword{internal}
