% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-roc.R
\name{print.srp.roc}
\alias{print.srp.roc}
\title{Print class of 'srp.roc'}
\usage{
\method{print}{srp.roc}(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{passed to print.}
}
\description{
Print class of 'srp.roc'
}
\keyword{internal}
