% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-roc.R
\name{roc}
\alias{roc}
\title{Build ROCs}
\usage{
roc(
  data,
  outcome,
  exposure,
  positive = NULL,
  combine = FALSE,
  combine.only = FALSE,
  ci = TRUE,
  ci.method = c("delong", "bootstrap"),
  smooth = FALSE,
  smooth.args = list(),
  digits = 2
)
}
\arguments{
\item{data}{a data frame.}

\item{outcome}{the outcome variable name contain in data, which can be a factor
or a numerical variable, but must be a binary variable.}

\item{exposure}{one or more exposure variable names are included in data.}

\item{positive}{in which  positive of outcome variable to make the comparison.
By default, positive is automatically defined. If outcome is a factor variable,
then positive is defined as the highest level. If outcome is a numerical
variable, then positive is defined as the largest value.}

\item{combine}{a logical, indicating whether to use multiple exposure variables
for joint prediction using logistic regression.}

\item{combine.only}{a logical, indicating whether to show only the results of
joint predictions.}

\item{ci}{a logical, indicating whether to show ci.}

\item{ci.method}{the method to use, either “delong” or “bootstrap”. The
first letter is sufficient. If omitted, the appropriate method is selected as
explained in details.}

\item{smooth}{a logical, indicating whether to smooth a ROC curve.}

\item{smooth.args}{further arguments for smooth, see \link[pROC:smooth]{pROC::smooth} function.}

\item{digits}{digits, default 2.}
}
\value{
a data frame.
}
\description{
Build ROCs
}
