% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-select.R
\name{select_numeric}
\alias{select_numeric}
\alias{select_factor}
\alias{select_category}
\alias{select_logical}
\alias{select_variable}
\title{Select variables}
\usage{
select_numeric(data, varnames = NULL, type = c("name", "data", "index"))

select_factor(data, varnames = NULL, type = c("name", "data", "index"))

select_category(data, varnames = NULL, type = c("name", "data", "index"))

select_logical(data, varnames = NULL, type = c("name", "data", "index"))

select_variable(data, ..., type = c("name", "data", "index"))
}
\arguments{
\item{data}{a data frame.}

\item{varnames}{variable names，Indicates to select the corresponding data
type in the specified variable, if it is NULL, it means to select from all
variables.}

\item{type}{return type, 'name' for a vector of character, 'index' for a
vector of integer, and 'data' for a data frame.}

\item{...}{variables to select.If not specified, all variables will be returned.}
}
\value{
depends on the parameter type. 'name' for a vector of character,
'index' for a vector of integer, and 'data' for a data frame.
}
\description{
Select variables in a data frame.
}
