% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-normality.R
\name{shapiro_wilk}
\alias{shapiro_wilk}
\alias{ks_test}
\alias{kolmogorov_smirnov}
\alias{shapiro_francia}
\title{Normality test method}
\usage{
shapiro_wilk(
  data,
  group = NULL,
  varnames = NULL,
  language = NULL,
  labels = NULL,
  table.number = NULL,
  digits = 3,
  ...
)

ks_test(
  data,
  group = NULL,
  varnames = NULL,
  language = NULL,
  labels = NULL,
  table.number = NULL,
  digits = 3,
  ...
)

kolmogorov_smirnov(
  data,
  group = NULL,
  varnames = NULL,
  language = NULL,
  labels = NULL,
  table.number = NULL,
  digits = 3,
  ...
)

shapiro_francia(
  data,
  group = NULL,
  varnames = NULL,
  language = NULL,
  labels = NULL,
  table.number = NULL,
  digits = 3,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{group}{group variable name.}

\item{varnames}{numeric variable names.}

\item{language}{language language, typically “en”, or "zh", default "en".}

\item{labels}{a data frame contain the column of term, code, and label.}

\item{table.number}{table number.}

\item{digits}{digits for p value.}

\item{...}{unused.}
}
\value{
a data frame with class of 'srp.norm'
}
\description{
Normality test method
}
