% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{srpubr-package}
\alias{srpubr}
\alias{srpubr-package}
\title{srpubr: Statistical Report for Publication Ready}
\description{
Statistical Report for Publication Ready.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kunhuo/srpubr}
}

}
\keyword{internal}
