% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-missing-pattern.R
\name{gg_missing_pattern}
\alias{gg_missing_pattern}
\title{Plot missing pattern with ggplot2}
\usage{
gg_missing_pattern(
  data,
  show.all = FALSE,
  decreasing = TRUE,
  language = c("en", "zh"),
  font.family = "serif",
  font.size = 12,
  color = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{decreasing}{a logical. Should the sort order be increasing or decreasing?
if is NULL, not sorted.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{font.family}{font family.}

\item{font.size}{font size.}

\item{color}{fill color, must have two values, the first for missing, the second for observed.}

\item{...}{further arguments pass to \link{gg_theme_sci} function.}
}
\value{
a ggplot.
}
\description{
Plot missing pattern with ggplot2
}
\examples{
# Basic example
gg_missing_pattern(lung)

# show all variables
gg_missing_pattern(lung, show.all = TRUE)

# Specify the color
gg_missing_pattern(lung, show.all = TRUE, color = c("#E85827", "#81D8D0"))
}
