% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-missing.R
\name{identify_missing}
\alias{identify_missing}
\title{Missing value analysis}
\usage{
identify_missing(
  data,
  show.all = FALSE,
  decreasing = TRUE,
  detail.type = TRUE,
  digits = 1,
  language = c("en", "zh"),
  table.number = NULL,
  ...
)
}
\arguments{
\item{data}{a data frame.}

\item{show.all}{a logical, indicate whether to show all variables, by default,
only show missing variables.}

\item{decreasing}{a logical. Should the sort order be increasing or decreasing?
if is NULL, not sorted.}

\item{detail.type}{detail type.}

\item{digits}{digits for missing percent, defualt 2.}

\item{language}{language, typically “en”, or "zh", default "en".}

\item{table.number}{table number.}

\item{...}{unused arguments.}
}
\value{
a data frame.
}
\description{
Missing value analysis
}
\examples{
# Basic example
identify_missing(lung)

# Show all variable
identify_missing(lung, show.all = TRUE)

# increasing by missing count
identify_missing(lung, show.all = TRUE, decreasing = FALSE)
}
