% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{Overview a data frame}
\usage{
overview(data, digits = 3, ...)
}
\arguments{
\item{data}{a data frame.}

\item{digits}{digits, default 2.}

\item{...}{unused.}
}
\value{
a data frame.
}
\description{
Overview a data frame
}
\examples{
overview(aSAH)

overview(lung)
}
